IOS-3386 Initial TX builder, fee estimation, getting account info in wallet manager

+++ BlockchainSdk/WalletManagers/Cosmos/CosmosModels.swift
@@ -61,7 +61,7 @@ extension CosmosTxResponse {
 // MARK: - Local models
 
 struct CosmosAccountInfo {
-    let accountNumber: Int
-    let sequenceNumber: Int
+    let accountNumber: UInt64
+    let sequenceNumber: UInt64
     let amount: Amount
 }
